Gather all TransitGateway resources
